Refactored and added to examples
